chore: 内置tree折叠状态
