+++ packages/graphs/src/utils/index.ts
@@ -7,7 +7,7 @@ export { getStatusCfg } from './get-status-cfg';
 export { setEllipsis } from './set-ellipsis';
 export { getChildrenData } from './get-children-data';
 export { getStyle } from './get-style';
-export { bindDefaultEvents, bindRadialExplore, bindStateEvents, bindSourceMapCollapseEvents } from './events';
+export { bindStateEvents } from './state-events';
 export { deepClone } from './deep-clone';
 export { getAnchorPoints } from './get-anchor-points';
 export { getCommonCfg } from './get-common-cfg';
