+++ packages/site/examples/relation-graph/flow-analysis-graph/demo/custom.js
@@ -205,7 +205,6 @@ const DemoFlowAnalysisGraph = () => {
       return {
         position: 'right',
         show: edges.find((item) => item.source === cfg.id),
-        collapsed: !edges.find((item) => item.source === cfg.id),
       };
     },
     layout: {
