+++ packages/site/examples/relation-graph/fund-flow-graph/demo/style.js
@@ -124,7 +124,6 @@ const DemoFundFlowGraph = () => {
       return {
         position: 'right',
         show: edges.find((item) => item.source === cfg.id),
-        collapsed: !edges.find((item) => item.source === cfg.id),
       };
     },
   };
